requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyJWT==0.4.3
Werkzeug==0.10.1
coverage==4.0a5
itsdangerous==0.24
mongokit==0.9.1.1
nose==1.3.4
pymongo==2.8
wsgiref==0.1.2
